Rename modules and package names
The modules are now called lib and app, and all capital letters have
been taken out package names (well, directories only, the packages were
already lowercase).
