+++ app/src/main/res/drawable-hdpi/ic_in_call_touch_handle_normal.png
