+++ app/src/main/res/drawable-hdpi/ic_lockscreen_answer_normal.png
