+++ app/src/main/res/drawable-hdpi/ic_lockscreen_decline_normal.png
