+++ app/src/main/res/drawable-hdpi/ic_lockscreen_glowdot.png
