+++ app/src/main/res/drawable-mdpi/ic_in_call_touch_handle_normal.png
