+++ app/src/main/res/drawable-mdpi/ic_lockscreen_answer_activated.png
