+++ app/src/main/res/drawable-mdpi/ic_lockscreen_decline_activated.png
