+++ app/src/main/res/drawable-xhdpi/ic_in_call_touch_handle_normal.png
