+++ app/src/main/res/drawable-xhdpi/ic_lockscreen_answer_normal.png
