+++ app/src/main/res/drawable-xhdpi/ic_lockscreen_decline_activated.png
