+++ app/src/main/res/drawable/ic_in_call_touch_handle.xml
