+++ app/src/main/res/drawable/ic_lockscreen_answer.xml
