+++ app/src/main/res/drawable/ic_lockscreen_answer_activated_layer.xml
