+++ app/src/main/res/drawable/ic_lockscreen_answer_normal_layer.xml
