+++ app/src/main/res/drawable/ic_lockscreen_decline_activated_layer.xml
