+++ app/src/main/res/drawable/ic_lockscreen_decline_normal_layer.xml
