+++ app/src/main/res/drawable/ic_lockscreen_outerring.xml
