+++ app/src/main/res/values/arrays.xml
