+++ lib/gradle.properties
