+++ lib/proguard-rules.txt
