+++ lib/src/main/java/net/frakbot/glowpadbackport/Ease.java
