+++ lib/src/main/java/net/frakbot/glowpadbackport/util/Const.java
