+++ lib/src/main/java/net/frakbot/glowpadbackport/util/TimeInterpolator.java
